Finished homepage both mobile and desktop view
